Use string Hash indexes in put_object instead of symbols
The `{'foo': 'bar'}` syntax represents a Hash with index `:foo`. Calls to merge that Hash with `{'foo' => 'bar'}` result in a Hash with two distinct elements instead of the original value being overridden. This manifested itself in `put_object` as the `content_type` option being ignored. There was also a duplicate `Authorization` header sent.