Add prefix for call to files in a bucket
Limit calls to :files in a ::Directory instance to return files with a
prefix in the bucket. As opposed to returning all files in a bucket
under all of its folders as it previously would do with :new on ::Files
instances.

A prefix should be given as a keyword argument to ::Directories
collections as such in service.directories.get('bucket', prefix: prefix)

At ::Directories.get, non-existent buckets will raise an exception if 
they can't be found